Data type cleanup (FileSystem classes)

+++ Emulator/FileSystems/FSBitmapBlock.cpp
@@ -40,7 +40,7 @@ void
 FSBitmapBlock::dump() const
 {
     u32 count = 0;
-    for (u32 i = 1; i < bsize() / 4; i++) {
+    for (isize i = 1; i < bsize() / 4; i++) {
         if (u32 value = get32(i)) {
             for (isize j = 0; j < 32; j++) {
                 if (GET_BIT(value, j)) count++;
